requirements.txt
Flask==3.0.3
numpy
pandas==2.2.2
scikit_learn==1.4.2
torch --index-url https://download.pytorch.org/whl/cpu
transformers
langchain_community
ollama
